enrollment-notes: add NoteModelBuilder; adapt NoteHandler; refactor imports
