+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/TeisEndPointCall.java
@@ -10,7 +10,7 @@ import org.hisp.dhis.android.core.data.api.Fields;
 import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
 import org.hisp.dhis.android.core.data.database.Transaction;
 import org.hisp.dhis.android.core.enrollment.Enrollment;
-import org.hisp.dhis.android.core.enrollment.Note.Note;
+import org.hisp.dhis.android.core.enrollment.note.Note;
 import org.hisp.dhis.android.core.event.Event;
 import org.hisp.dhis.android.core.relationship.Relationship;
 import org.hisp.dhis.android.core.resource.ResourceHandler;
@@ -20,7 +20,6 @@ import java.util.Date;
 import java.util.List;
 
 import retrofit2.Response;
-
 
 public class TeisEndPointCall implements Call<Response<Payload<TrackedEntityInstance>>> {
